+++ runtime/src/lib.rs
@@ -11,7 +11,6 @@ pub mod bank_forks;
 pub mod bank_utils;
 pub mod commitment;
 pub mod compute_budget_details;
-mod epoch_rewards_hasher;
 pub mod epoch_stakes;
 pub mod genesis_utils;
 pub mod installed_scheduler_pool;
